Refactor backends to inherit from ComponentBasedBackend
